memory db
graph map to binary search
$aA and $aK
encrypt / decrypt props
txn paths
writeDo to writeAction
AceGraph types
wal to aol
add collator to memory

+++ src/schema/revertWriteSchema.js
@@ -1,26 +1,21 @@
-import { td } from '#ace'
 import { Memory } from '../objects/Memory.js'
 import { unlink, writeFile } from 'node:fs/promises'
-import { doesPathExist, getPaths, initPaths } from '../util/file.js'
+import { doesPathExist, initPaths } from '../util/file.js'
 
 
-/**
- * @param { td.AceFnOptions } options 
- * @returns { Promise<void> }
- */
-export async function revertWriteSchema (options) {
-  if (Memory.txn.env) {
+/** @returns { Promise<void> } */
+export async function revertWriteSchema () {
+  if (Memory.txn.env && Memory.txn.paths) {
     let nowLastId = Memory.txn.schemaNowDetails?.[Memory.txn.env]?.lastId
     const originalLastId = Memory.txn.schemaOriginalDetails?.[Memory.txn.env]?.lastId
 
     if (typeof nowLastId === 'number' && typeof originalLastId === 'number' && nowLastId > originalLastId) { // if we created some schemas since this txn began
-      const paths = getPaths(options.dir, ['dir', 'schemas', 'schemaDetails'])
 
-      await initPaths(paths, ['dir', 'schemas'])
-      await writeFile(paths.schemaDetails, JSON.stringify(Memory.txn.schemaOriginalDetails), 'utf-8')
+      await initPaths(['dir', 'schemas'])
+      await writeFile(Memory.txn.paths.schemaDetails, JSON.stringify(Memory.txn.schemaOriginalDetails), 'utf-8')
 
       while (nowLastId > originalLastId) {
-        const schemaPath = `${ paths.schemas }/${ nowLastId }.json`
+        const schemaPath = `${ Memory.txn.paths.schemas }/${ nowLastId }.json`
         if (await doesPathExist(schemaPath)) await unlink(schemaPath)
         nowLastId--
       }
